feat: 添加 NodeDebugResponse 组件以增强调试功能
